+++ src/methods/getEntry.js
@@ -1,0 +1,15 @@
+export default ({ production, entry, vendor}) => {
+  // hmr enable via env
+  !production && vendor.unshift('webpack-hot-middleware/client');
+
+  if (vendor.length) {
+    return {
+      main: entry,
+      vendor: vendor
+    }
+  } else {
+    return {
+      index: entry
+    }
+  }
+}