+++ HISTORY.md
@@ -2,6 +2,17 @@
 
 ---
 
+## 0.2.3
+
+Minor modify.
+
+## 0.2.2
+
+Adapt Relay projects
+News:
+  * argument `tapas` has a new key `babelLoaderPlugin`, because of babelrc [issue](https://phabricator.babeljs.io/T2529)
+  * argument `tapas` has a new key `port`, you can set your own port to server webpack
+
 ## 0.2.0
 
 First usable version.
@@ -9,3 +20,5 @@ First usable version.
 ## 0.1.0
 
 Basic requirement is implemented.
+
+